Remove single use FunctionToImmutableConfigBase
No reason to have a baseclass if we only really use it once.
